........
  r1326949 | ay | 2012-04-17 08:54:08 +0200 (Tue, 17 Apr 2012) | 1 line
  
  [CXF-4191] RM broken in synchronous Mode
........


git-svn-id: https://svn.apache.org/repos/asf/cxf/branches/2.5.x-fixes@1327040 13f79535-47bb-0310-9956-ffa450edef68

+++ rt/ws/rm/src/main/java/org/apache/cxf/ws/rm/SourceSequence.java
@@ -144,10 +144,17 @@ public class SourceSequence extends AbstractSequence {
         acknowledgement = a;
         source.getManager().getRetransmissionQueue().purgeAcknowledged(this);
         if (allAcknowledged()) {
-            RMEndpoint rme = source.getReliableEndpoint();
-            Proxy proxy = rme.getProxy();
-            proxy.terminate(this);
-            source.removeSequence(this);
+            if (null == target || RMUtils.getAddressingConstants().getAnonymousURI().equals(
+                target.getAddress().getValue())) {
+                LOG.log(Level.WARNING, "STANDALONE_ANON_TERMINATE_SEQUENCE_MSG");
+                // keep the sequence and let RMOutInterceptor remove it after building the TS message
+                // if we remove the sequence here, RMOutInterceptor should check for a null sequence
+            } else {
+                RMEndpoint rme = source.getReliableEndpoint();
+                Proxy proxy = rme.getProxy();
+                proxy.terminate(this);
+                source.removeSequence(this);
+            }
         }
     }
